Handle WAV metadata properly
The initial implementation, made by ftrvxmtrx (https://ftrv.se/_about)
assumed that metadata related to the author of the song are set before the DATA keyword.
Unfortunately, this assumption is false. Indeed, the specification of RIIF does not specify
where the chunk related to those metadata (the INFO tag) must be placed.
Consequently, the chunk can be placed before or after the data section.

This commit handles the INFO chunk without making assumptions about where this section is.
It will also retrieve metadata properly.

Links:

- RIFF (INFO chunk placement problems): https://en.wikipedia.org/wiki/Resource_Interchange_File_Format#INFO_chunk_placement_problems